requirements.txt
python-dotenv>=1.0.0
ollama>=0.3.0
Pillow>=10.0.0
pypdf>=4.0.0
edge-tts>=6.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0
# Optional anchor TTS (GPU): pip install git+https://github.com/suno-ai/bark.git
# bark
